prevent deletion of the `rauthy` client directly on the API endpoint

+++ rauthy-handlers/src/clients.rs
@@ -400,6 +400,7 @@ pub async fn get_client_logo(
             Logo::find_cached(&data, "rauthy", &LogoType::Client).await?
         }
     };
+
     Ok(HttpResponse::Ok()
         .insert_header((CONTENT_TYPE, logo.content_type))
         .body(logo.data))
@@ -543,12 +544,21 @@ pub async fn put_generate_client_secret(
 #[delete("/clients/{id}")]
 pub async fn delete_client(
     data: web::Data<AppState>,
-    path: web::Path<String>,
+    id: web::Path<String>,
     principal: ReqPrincipal,
 ) -> Result<HttpResponse, ErrorResponse> {
     principal.validate_api_key_or_admin_session(AccessGroup::Clients, AccessRights::Delete)?;
 
-    let client = Client::find(&data, path.into_inner()).await?;
+    let id = id.into_inner();
+
+    if &id == "rauthy" {
+        return Err(ErrorResponse::new(
+            ErrorResponseType::BadRequest,
+            "The `rauthy` client must not be deleted".to_string(),
+        ));
+    }
+
+    let client = Client::find(&data, id).await?;
     client.delete(&data).await?;
     Ok(HttpResponse::Ok().finish())
 }
